#944: Some refactoring
